test: share db fixture
